Log now has smart times...Finally
Those of you that know me…(Tommy and Nicky) know that I hate to be that
guy… but the log now should finally be more user friendly.  If you
don’t think its very user friendly then maybe its not the log.  Maybe
your just not a very friendly person!  You really need to work on that.
 To be honest it is the worst characteristic about you.  Once again I
hate to be the one who tells you this.  This commit is detected to
Tommy who wanted a laugh…Trying to make my commits more interesting to
read.  Sorry it's so long.  This is really funny,  this commit is so
stupid.  I really have too much time, but I just can’t stop writing.
Well, until the next commit…Brett